Moved program minimising transformation to new file.
